+++ cloudy_salesforce/__init__.py
@@ -10,6 +10,7 @@ from cloudy_salesforce.collections import DmlResult, delete, insert, update, ups
 from cloudy_salesforce.exceptions import SalesforceError
 from cloudy_salesforce.query import SoqlQuery, query, select, soql_query
 from cloudy_salesforce.sobjects import SObjects, parse_sobject_response, sobject
+from cloudy_salesforce.types import UNSET, UnsetType
 
 __version__ = "0.3.0"
 
@@ -31,5 +32,7 @@ __all__ = [
     "sobject",
     "SObjects",
     "parse_sobject_response",
+    "UNSET",
+    "UnsetType",
     "__version__",
 ]